Fix for DOCS-29: Add capability to generate PDF for a single chapter
The following example generates PDF only for the chapter whose file is named chap-one.xml:

mvn -DdocumentSrcName=chap-one.xml -Dinclude=pdf clean pre-site

I have not tried to do a foolproof implementation for this. If you have two chap-one.xml files in your sources, this might not work.
This also does not work if you want output for a section smaller than a file.

But it is often much faster than doing an entire pre-site build.